requirements.txt
PyMuPDF==1.22.5
langchain==0.0.267
openai==0.27.8
streamlit==1.25.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
